backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.8.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
